refactor: stomp client를 contextAPI로 관리하도록 변경
